// src/film/schema/types/film.type.ts
import { BaseModel } from '../../../common/schema/base-mode.type';

export interface Film extends BaseModel {
  filmId: string;
  externalId?: string;
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  species: string;
  starships: string;
  vehicles: string;
  characters: string;
  planets: string;
  url: string;
  created: string;
  edited: string;
}